requirements.txt
Flask
Flask-WTF
undetected-playwright-patch
playwright==1.40.0